requirements.txt
google-generativeai
pyautogui
pyttsx3
SpeechRecognition
websockets
PyAudio
opencv-python
